move api logic to proto files
