+++ src/objects/client_proto/index.js
@@ -1,0 +1,4 @@
+module.exports = {
+    ChatProto: require('./ChatProto'),
+    SocketProto: require('./SocketProto')
+}
